[242201] Make it easier to switch between diagrams
Now, each time you open a new diagram in a given editor (by double clicking in the editor area or through the outline view), the new diagram is opened in the same editor. It is then possible to navigate between diagrams using the "previous", "parent" or "next" buttons. In fact, I have implemented a similar mechanism than the one available with editors done in the Topcased project.

Currently if you try to change the active diagram from the Project Explorer, the old mechanism is used (each diagram is opened in a new editor) : this still ensure a way to compare two diagrams visually.

There is a known (minor) bug on diagram navigation : indeed, when switching between two diagrams, the editor is marked as dirty, although none modification has been done neither on the domain model nor on the notational one.