456188: EReferences listed within the EClass
Introduce a specific subnode mapping within the EClass one to display
EReferences which are contained or inherited (in italic then) and whose
target is not part of the diagram.
Such text rendering for such EReferences obey the same rules as all the
other elements, [1..1] cardinality bring bold font, and direct edit also
provide the exact same shortcuts as editing an EAttribute.

This mapping is synchronized and enabled by default (and not in a
specific layer) though can be disabled by the user through the filter
"Hide references (nodes)".


Bug: 456188
Change-Id: I7c0363ea45bc7857e5106e8ee17dfb9f25c04045
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=456188