[495102] Streamline the "Initialize Ecore Diagram" procedure
This commit:
a°) rephrase the representation labels to make them easier to understand
b°) replace the empty diagram notice by a less verbose one and stop
displaying it when there is no content in the EPackage to add in the
diagram.
c°) replace the calls to several dialogs and wizards with an unified
workflow when clicking on "Initialize Ecore Diagram"

This unified wizard make sure one can always pick a different .aird for
each call except when we are in the context of a modeling project (which
does not allow this) and also present the type of representations
without even mentioning the Viewpoints. Also the class diagram is
pre-selected and the first semantic element which is matching the
diagram type is also pre-selected, leading to a short Next, Next, Finish
behavior.

Bug: 495102
Change-Id: Ia7c23caed13ce9d5189ccc69e8c1fbca4a9cfd39