Remove EGenericTypes from semanticElements references
This commit change the VSM structure so that the ESuperType edge
mappings is no longer specifying that a reference to the EGenericTypes
of the EClass should be kept. As the genericsupertype reference in Ecore
has a fairly weird status (mutually derived, but not structurally
derived) that could cause a number of problems later one.

This commit replace this reference by a specific handling of the
properties views.

Change-Id: I92aa56527447eed90604122d82a3645e999d1e50