[571427] Replace override with a more direct conditional widget
The previous solution (in 1dcea7903c636a2e8962671d0f7e748a2a10ed9f)
used an override to add a new widget inside the "default_rules" when
on an EClassifier. This works for EClassifiers, but for all other
kinds of elements on which "default_rules" is also used, the override
produces a completely empty page.

Instead, use a dynamic if directly inside "default_rules" to
contribute the widget when on an EClassifier without side-effects on
other types of elements.

"Dynamic If" can not be used directly inside the group definition, but
this can be worked around using a "Dyanmic For" which iterates on a
single value (self).

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=571427
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>